Move HealthLogCache to Utils
