+++ GWToolboxdll/Utils/HealthLogCache.h
